use query file from qa.commons
Use a query template that is defined in the Qanary commons package,
and remove the query template that was locally defined for this component.
Change the binding variables to fit the new template.
Add a unit test for setLanguageText() to test the creation of INSERT
queries and updates to the triplestore.